# config/swagger.json
{
    "openapi": "3.0.0",
    "info": {
      "title": "Habit & Journal Tracker API",
      "version": "1.0.0",
      "description": "API for managing journal entries and habits"
    },
    "servers": [
      {
        "url": "https://habit-back.navong.cloud"
      }
    ],
    "paths": {
      "/api/habits": {
        "get": {
          "summary": "Retrieve all habits",
          "responses": {
            "200": {
              "description": "A list of habits"
            }
          }
        },
        "post": {
          "summary": "Create a new habit",
          "requestBody": {
            "required": true,
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "name": {
                      "type": "string"
                    },
                    "description": {
                      "type": "string"
                    },
                    "goal": {
                      "type": "string"
                    },
                    "frequency": {
                      "type": "string"
                    },
                    "color": {
                      "type": "string"
                    },
                    "category": {
                      "type": "string"
                    }
                  }
                }
              }
            }
          },
          "responses": {
            "201": {
              "description": "Habit created"
            }
          }
        },
        "put": {
          "summary": "Update a habit",
          "requestBody": {
            "required": true,
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "name": {
                      "type": "string"
                    },
                    "description": {
                      "type": "string"
                    },
                    "goal": {
                      "type": "string"
                    },
                    "frequency": {
                      "type": "string"
                    },
                    "color": {
                      "type": "string"
                    },
                    "category": {
                      "type": "string"
                    }
                  }
                }
              }
            }
          },
          "responses": {
            "200": {
              "description": "Habit updated"
            }
          }
        },
        "delete": {
          "summary": "Delete a habit",
          "responses": {
            "200": {
              "description": "Habit deleted"
            }
          }
        }
      }
    }
  }
  